backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "fuzzywuzzy>=0.18.0",
    "langchain-community>=0.3.31",
    "langchain-google-genai>=2.1.12",
    "langchain-pinecone>=0.2.12",
    "langchain-text-splitters>=0.3.11",
    "langgraph>=0.6.10",
    "pymupdf>=1.26.5",
    "python-levenshtein>=0.27.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
]
